Move exception to exceptions folder/namespace
PR feedback
