+++ src/ManageCourses.Api/Services/UserService.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using GovUk.Education.ManageCourses.Api.Exceptions;
 using GovUk.Education.ManageCourses.Api.Middleware;
 using GovUk.Education.ManageCourses.Api.Services.Email;
 using GovUk.Education.ManageCourses.Api.Services.Email.Model;
